move util classes from computer command
